requirements.txt
asgiref==3.7.2
attrs==23.1.0
crispy-bootstrap5==0.7
Django==4.0.10
django-ckeditor==6.3.2
django-cors-headers==4.1.0
django-crispy-forms==2.0
django-js-asset==2.0.0
django-resized==1.0.2
djangorestframework==3.13.1
drf-spectacular==0.22.1
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.17.3
Pillow==9.1.1
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
typing-extensions==4.6.3
uritemplate==4.1.1
django-robots
django-modeltranslation
scipy
nltk
pandas
